Correction rust

+++ lib/solver/src/model.rs
@@ -2,7 +2,7 @@ pub mod solver {
     use std::collections::LinkedList;
 
     pub trait Solvable {
-        fn value(&self)->f64;
+        fn value(&self)->i64;
         fn next_states(&self)->LinkedList< Self > where Self: Sized;
     }
